+++ frameworks/native/ipc/src/user_auth_callback_proxy.cpp
@@ -45,7 +45,7 @@ void UserAuthCallbackProxy::OnResult(int32_t result, const Attributes &extraInfo
         return;
     }
 
-    bool ret = SendRequest(static_cast<uint32_t>(UserAuthInterfaceCode::USER_AUTH_ON_RESULT), data, reply);
+    bool ret = SendRequest(UserAuthInterfaceCode::USER_AUTH_ON_RESULT, data, reply);
     if (!ret) {
         IAM_LOGE("send request failed");
     }
@@ -76,7 +76,7 @@ void UserAuthCallbackProxy::OnAcquireInfo(int32_t module, int32_t acquireInfo, c
         return;
     }
 
-    bool ret = SendRequest(static_cast<uint32_t>(UserAuthInterfaceCode::USER_AUTH_ACQUIRE_INFO), data, reply);
+    bool ret = SendRequest(UserAuthInterfaceCode::USER_AUTH_ACQUIRE_INFO, data, reply);
     if (!ret) {
         IAM_LOGE("send request failed");
     }
@@ -121,7 +121,7 @@ void GetExecutorPropertyCallbackProxy::OnGetExecutorPropertyResult(int32_t resul
         return;
     }
 
-    bool ret = SendRequest(static_cast<uint32_t>(UserAuthInterfaceCode::USER_AUTH_GET_EX_PROP), data, reply);
+    bool ret = SendRequest(UserAuthInterfaceCode::USER_AUTH_GET_EX_PROP, data, reply);
     if (!ret) {
         IAM_LOGE("send request failed");
     }
@@ -161,7 +161,7 @@ void SetExecutorPropertyCallbackProxy::OnSetExecutorPropertyResult(int32_t resul
         return;
     }
 
-    bool ret = SendRequest(static_cast<uint32_t>(UserAuthInterfaceCode::USER_AUTH_SET_EX_PROP), data, reply);
+    bool ret = SendRequest(UserAuthInterfaceCode::USER_AUTH_SET_EX_PROP, data, reply);
     if (!ret) {
         IAM_LOGE("send request failed");
     }
